style: Auto-fix 1649 ruff violations (79% reduction)
Ruff auto-fix results:
- Fixed 882 issues with --fix
- Fixed 767 additional issues with --unsafe-fixes
- Total fixes: 1649 errors
- Remaining: 422 errors (from 2041)

Main fixes applied:
- D400/D415: Added periods to docstring first lines
- D413: Added blank lines after docstring sections
- RUF100: Removed unused noqa directives
- RUF010: Used explicit conversion flags in f-strings
- TC001/TC003: Moved type-only imports to TYPE_CHECKING
- UP037: Removed quotes from type annotations

Remaining issues (non-auto-fixable):
- 214 D401 (non-imperative mood) - docstring style
- 123 D205 (missing blank line) - docstring format
- 42 RUF012 (mutable class default) - design issues
- 32 RUF001 (ambiguous unicode) - Greek characters
- 8 D417 (undocumented param) - missing docs
- 1 RUF034 (useless if-else) - logic simplification

Files modified: 382
Lines: +1562 / -829

+++ oncutf/ui/widgets/node_editor/graphics/scene.py
@@ -43,6 +43,7 @@ class QDMGraphicsScene(QGraphicsScene):
         scene: Reference to logical Scene model.
         gridSize: Pixel size of grid cells.
         gridSquares: Number of cells between major (dark) grid lines.
+
     """
 
     item_selected = Signal()
@@ -54,6 +55,7 @@ class QDMGraphicsScene(QGraphicsScene):
         Args:
             scene: Logical Scene this graphics scene represents.
             parent: Optional parent widget.
+
         """
         super().__init__(parent)
 
@@ -85,6 +87,7 @@ class QDMGraphicsScene(QGraphicsScene):
 
         Args:
             event: Qt drag move event.
+
         """
 
     def set_graphics_scene_rect(self, width: int, height: int) -> None:
@@ -93,6 +96,7 @@ class QDMGraphicsScene(QGraphicsScene):
         Args:
             width: Total scene width in pixels.
             height: Total scene height in pixels.
+
         """
         self.setSceneRect(-width // 2, -height // 2, width, height)
 
@@ -105,13 +109,14 @@ class QDMGraphicsScene(QGraphicsScene):
         Args:
             painter: QPainter for drawing operations.
             rect: Visible rectangle area to draw.
+
         """
         super().drawBackground(painter, rect)
 
-        left = int(math.floor(rect.left()))
-        right = int(math.ceil(rect.right()))
-        top = int(math.floor(rect.top()))
-        bottom = int(math.ceil(rect.bottom()))
+        left = math.floor(rect.left())
+        right = math.ceil(rect.right())
+        top = math.floor(rect.top())
+        bottom = math.ceil(rect.bottom())
 
         first_left = left - (left % self.gridSize)
         first_top = top - (top % self.gridSize)
